fix: added struct tags for json and changed Amount int->int64

+++ internal/sqlite/sqlite_repository.go
@@ -24,7 +24,7 @@ type dbExpense struct {
 	CreatedAt   int64
 	OccuredAt   int64
 	Description string
-	Amount      int
+	Amount      int64
 }
 
 func toDBExpense(e *expenses.Expense) dbExpense {
